+++ shim/shim-rust/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["shim-core", "shim", "shim-macos"]
+members = ["shim", "shim-macos"]
 resolver = "2"
