--- shim/shim-rust/shim-core/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "shim-core"
-version = "1.1.0"
-edition = "2021"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-semver = "1.0"
-directories = "5.0"
