+++ shim/shim-rust/shim/Cargo.toml
@@ -4,7 +4,10 @@ version = "1.1.0"
 edition = "2021"
 
 [dependencies]
-shim-core = { path = "../shim-core" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+semver = "1.0"
+directories = "5.0"
 
 [build-dependencies]
 winres = "0.1"
